fix: resolve Edge Runtime createContext error in middleware
- Split ROUTES constant into separate routes.ts file without React dependencies
- Move React-dependent navigation components to navigation.tsx
- Update proxy.ts to import from edge-compatible routes.ts
- Fix middleware bundling issue causing 'createContext is not a function' error in Docker

This resolves the TypeError that occurred when middleware tried to use
React's createContext in Edge Runtime environment.
