+++ src/utils/constants/index.ts
@@ -5,4 +5,5 @@
 export * from './theme';
 export * from './http';
 export * from './homepage';
-export * from './navigation';
+export * from './routes'; // Edge runtime compatible routes
+export * from './navigation'; // Client-side navigation with React components
